Add submit on all steps
